Backtrack schema changes to version 7
